+++ bootstrap/app.php
@@ -84,6 +84,7 @@ return Application::configure(basePath: dirname(__DIR__))
 
         // Отправка запросов в ИИ
         $schedule->command('app:ai:health-check')->everyFifteenMinutes()->withoutOverlapping();
+        $schedule->command('app:ai:yandex-health-check')->everyFifteenMinutes()->withoutOverlapping();
 
         $schedule->everyThirtyMinutes()
             ->withoutOverlapping()